pyproject.toml
[tool.poetry]
name = "main_service"
version = "0.1.0"
description = ""
authors = ["Jacob Zuliani <jake@burla.dev>"]
packages = [{include = "main_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
jsonschema = "^4.17.3"
docker = "^6.1.3"
google-cloud-logging = "^3.5.0"
grpcio = "^1.54.2"
pyyaml = "^6.0"
google-cloud-firestore = "^2.11.1"
google-cloud-secret-manager = "^2.16.2"
slack-sdk = "^3.21.3"
google-cloud-batch = "^0.14.0"
google-cloud-pubsub = "^2.18.2"
google-cloud-run = "^0.9.1"
google-api-python-client = "^2.97.0"
google-auth = "^2.22.0"
cryptography = "^41.0.3"
google-cloud-storage = "^2.10.0"
cloudpickle = "^2.2.1"
tblib = "^3.0.0"
modal = "^0.55.4036"
google-cloud-compute = "^1.14.1"
paramiko = "^3.3.1"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
starlette = "^0.35.1"
itsdangerous = "^2.1.2"
httpx = "^0.26.0"
authlib = "^1.3.0"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
